on halfway to adapter refactoring, password toggle is fixed
