+++ app/build.gradle
@@ -70,8 +70,6 @@ android {
 
 dependencies {
     implementation 'com.squareup.moshi:moshi-kotlin:1.13.0'
-
-    implementation 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.3.2'
     // Import the BoM for the Firebase platform
     implementation platform('com.google.firebase:firebase-bom:29.0.1')
 
@@ -106,6 +104,7 @@ dependencies {
     implementation("androidx.navigation:navigation-dynamic-features-fragment:2.3.5")
 
     implementation 'androidx.core:core-ktx:1.7.0'
+    implementation 'com.android.support:design:25.0.0'
     implementation 'androidx.appcompat:appcompat:1.4.0'
     implementation 'com.google.android.material:material:1.4.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.2'
