+++ app/src/main/java/com/example/moodtrackerproject/data/DataBaseRepository.kt
@@ -1,6 +1,5 @@
 package com.example.moodtrackerproject.data
 
-import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import com.example.moodtrackerproject.domain.NoteBody
 import com.example.moodtrackerproject.utils.Preference
@@ -12,8 +11,8 @@ import java.lang.reflect.ParameterizedType
 
 // single source of truth
 object DataBaseRepository {
-    var allNotes = MutableLiveData<MutableList<NoteBody>>(mutableListOf())
-    var favoriteNotes = MutableLiveData<MutableList<NoteBody>>(mutableListOf())
+    val allNotes = MutableLiveData<MutableList<NoteBody>>(mutableListOf())
+    val favoriteNotes = MutableLiveData<MutableList<NoteBody>>(mutableListOf())
     private val moshi = Moshi.Builder().addLast(KotlinJsonAdapterFactory()).build()
     private val values: ParameterizedType =
         Types.newParameterizedType(List::class.java, NoteBody::class.java)
@@ -22,22 +21,20 @@ object DataBaseRepository {
     fun insert(noteBody: NoteBody, onSuccess: () -> Unit) {
 
         allNotes.value = Preference.getNotes()
-        Log.d("aaaaaaaaaaaaaaaaa", allNotes.value.toString())
         allNotes.value!!.add(noteBody)
-        Log.d("bbbbbbbbbbbbbbbb", allNotes.value.toString())
+        // allNotes.value?.let{.add(noteBody)}
         var notesList: List<NoteBody> = allNotes.value!!
         val serNotes = jsonAdapter.toJson(notesList)
         // here i have to concat new notes with previous
         Preference.setNotes(serNotes)
 
-
-        favoriteNotes = DataBaseRepository.favoriteNotes.value?.map {
-            if (it.noteId == noteId) {
-                it.copy(isChecked = !it.checked)
-            } else {
-                it
-            }
-        }
+//        favoriteNotes = DataBaseRepository.favoriteNotes.value?.map {
+//            if (it.noteId == noteId) {
+//                it.copy(isChecked = !it.checked)
+//            } else {
+//                it
+//            }
+//        }
     }
 
     fun delete(noteBody: NoteBody, onSuccess: () -> Unit) {
