+++ app/src/main/java/com/example/moodtrackerproject/ui/notes/details/NoteDetailsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.moodtrackerproject.ui.notes
+package com.example.moodtrackerproject.ui.notes.details
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.moodtrackerproject.R
 
-class InsideNoteFragment : Fragment() {
+class NoteDetailsFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_inside_note, container, false)
+        return inflater.inflate(R.layout.fragment_note_details, container, false)
     }
 }
