+++ app/src/main/java/com/example/moodtrackerproject/ui/notes/list/NotesAdapter.kt
@@ -1,20 +1,15 @@
 package com.example.moodtrackerproject.ui.notes.list
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.moodtrackerproject.R
-import com.example.moodtrackerproject.domain.NoteBody
 import com.example.moodtrackerproject.databinding.NoteItemBinding
-import kotlinx.android.synthetic.main.note_item.view.*
+import com.example.moodtrackerproject.domain.NoteBody
 
-class NotesAdapter(var listOfNotes: List<NoteBody> = emptyList()) : RecyclerView.Adapter<NotesAdapter.NotesHolder>() {
-
-//    private var isFavourite = false
+class NotesAdapter(var listOfNotes: List<NoteBodyUiModel> = emptyList()) : RecyclerView.Adapter<NotesAdapter.NotesHolder>() {
 
     class NotesHolder(private val binding: NoteItemBinding) : RecyclerView.ViewHolder(binding.root) {
-
         fun bind(model: NoteBodyUiModel) {
             binding.run {
                 root.setOnClickListener {
@@ -23,13 +18,13 @@ class NotesAdapter(var listOfNotes: List<NoteBody> = emptyList()) : RecyclerView
                 }
                 noteTextDateTime.text = model.date
                 noteTitleItem.text = model.title
-                noteText.text = model.text
+                noteTextItem.text = model.text
                 noteStarButtonUnchecked.setImageResource(
                     if (model.isChecked) R.drawable.ic_note_star_checked
                     else R.drawable.ic_note_star_unchecked
                 )
                 noteStarButtonUnchecked.setOnClickListener {
-                    model.checkChanged(noteId)
+                    model.checkChanged(model.noteId)
                 }
             }
         }
@@ -58,12 +53,11 @@ class NotesAdapter(var listOfNotes: List<NoteBody> = emptyList()) : RecyclerView
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): NotesHolder {
         return NotesHolder(
             NoteItemBinding.inflate(
-                LayoutInflater.from(this),
+                LayoutInflater.from(parent.context),
                 parent,
                 false
             )
         )
-
     }
 
     override fun onBindViewHolder(holder: NotesHolder, position: Int) {
@@ -73,8 +67,7 @@ class NotesAdapter(var listOfNotes: List<NoteBody> = emptyList()) : RecyclerView
     override fun getItemCount(): Int = listOfNotes.size
 
     fun setList(list: List<NoteBody>) {
-        listOfNotes = list
-        notifyDataSetChanged()
-        Log.d("ooooooooooooooooooooooooo", listOfNotes.toString())
+//        listOfNotes = list
+//        notifyDataSetChanged()
     }
 }
