+++ app/src/main/res-screen/login/layout/fragment_login_screen.xml
@@ -87,15 +87,15 @@
         android:hint="@string/password"
         android:textColorHint="@color/light_purple"
         app:boxStrokeColor="@color/light_purple"
-        app:endIconDrawable="@drawable/ic_welcome_lock"
-        app:endIconMode="custom"
-        app:errorIconTint="@color/light_purple"
+        app:passwordToggleEnabled="true"
+        app:passwordToggleTint="@color/black"
         app:layout_constraintEnd_toEndOf="@id/guideline_right"
         app:layout_constraintStart_toStartOf="@id/guideline_left"
         app:layout_constraintTop_toBottomOf="@id/email">
 
         <com.google.android.material.textfield.TextInputEditText
             android:id="@+id/pass_input"
+            android:inputType="textPassword"
             android:layout_width="match_parent"
             android:layout_height="wrap_content" />
