+++ app/src/main/res-screen/notes/layout/fragment_add_new_note.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.notes.AddNewNoteFragment">
+    tools:context=".ui.notes.add.AddNewNoteFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/toolbar_layout"
