+++ app/src/main/res-screen/notes/layout/fragment_note_details.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.notes.InsideNoteFragment">
+    tools:context=".ui.notes.NoteDetailsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
