+++ app/src/main/res/drawable/show_password_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/ic_welcome_lock" android:state_checked="true"/>
+    <item android:drawable="@drawable/ic_note_star_unchecked"/>
+</selector>
